Cache optimization and plotting enhancement (#45)
* Improved testing coverage

* Split calculation of relevant quantities into a separate file

* Smoothed Makie extension and testing

More general implementation of keyword arguments to allow for more precise handling of plots. Tests updated accordingly

* Plotting bug fixes to resolve scaling issues. ACF confidence intervals

* Optimized implementation of fitting procedure by introducing cache structs

To avoid per-evaluation allocations (of parameter vectors as well as intermediate buffers), an EvalCache struct was introduced which FCSModel now contains. Moreover, since a large portion of the evaluation was previously based on determining where in the parameter vector given parameters were stored, a layout struct was formed to store these results for previous use

* Update FCSModel usage throughout library and tests